cmd/flux: don't enforce max blob size here
Don't hardwire the max blob size in the flux content subcommand.
Instead, allow requests exceeding the system configured size to fail
at the target service.

Caveat: 'flux content store --dry-run' will now accept a value
that is too large.  I'm not sure this is much of a deal.  If it ever
is we can add code to read the broker attribute in that case.