test/kvs: check large KVS values/dirs
Try to store a value to the KVS larger than the RFC 10
max blob size of 1m.  In the current code this will succeed
not only because the max blob size has been increased, but
because the KVS still implements a cache in front of the content
cache.  If a "dirty" object in the KVS cache cannot be transferred
to the backing store (the content cache in this case) it simply
remains dirty in the KVS cache.  It's still a good test going
forward, as the KVS transitions to RFC 11 treeobj's.

Test 10,000 entries on one directory.
This *may* come close to exceeding the RFC 10 max blob size for the
directory object.

If LONGTEST, test 1,000,000 entries in one directory.
This definitely will exceed the R FC 10 max blob size fo the directory
object.