optparse: memoize optparse program fullname
The "fullname" of a program is different from the program_name in
the case of subcommands, where the fullname will be "cmd subcommand"
for use in usage output, error messages, etc.

Clean up code a bit by adding a function optparse_fullname() which
initializes p->fullname as needed and returns the result. This should
make code less confusing, at the expense of another member in the optarse
struct.