build: add -Wno-error=strict-aliasing
The koji buildfarm appends -Wall to CFLAGS as part of its %configure
macro.  This overrides the -Wno-strict-aliasing that we add in
common/libutil/Makefile.am to work around libev issues.
Since we have -Werror this becomes fatal and flux-core
can't be built in koji with the default %configure.

A workaround is to set CFLAGS to something else in the spec
file before calling %configure, however then that drops all the other
compiler flags that koji wants to add, such as -Wp,-D_FORTIFY_SOURCE=2.
We can manually add these flags into the spec file, but then we have
hardwired something that belongs elsewhere.

This patch adds -Wno-error=strict-aliasing to the Makefile.am,
after -Werror.  That should allow the build to succeed without the
above workaround, producing only non-fatal warnings in the output.