t/t2004-hydra: Do not list path to flux
Do not explicitly list path to flux binary in hydra test.
The sharness suite adds the path to the correct flux binary
to the front of PATH automatically, so this is not needed.

Also, by explicitly listing FLUX_BUILD_DIR, we point at the built
flux and break testing against an installed version of flux with
FLUX_TEST_INSTALLED_PATH