build: derive version from git describe
Instead of substituting the AC_INIT package and name by extracting
it from the META file, hardwire the name, and call directly out to
"git describe" to obtain the version.

As with the previous AC_INIT code definition, we use the m4_esyscmd()
macro to call git describe, processing the result through awk to trim
the newline.  We could eliminate the awk call with m4_esyscmd_s(),
but that would require a newer version of autoconf than is currently
supplied in RHEL 6, one of our major development platforms.

This affects the name of the tarball generated by 'make dist' and
eliminates the potential error of generating a release tarball from
something other than the exact tag for the release.

In man page generation, use PACKAGE_NAME instead of META_NAME,
and PACKAGE_VERSION instead of META_VERSION.  The man pages,
as before, include a footer with the package name and version:

    flux-core 0.1.0            01/27/2016            FLUX(1)

except now development man pages will be identified as such:

    flux-core 0.1.0-2-gbded    02/01/2016            FLUX(1)

One caveat (or maybe it is a feature?) is that the "version" for
development tarballs will not follow semantic versioning conventions
(http://semver.org/).