build: drop flux-libs.conf
Now that libraries are installed directly to $(libdir),
not $(libdir)/flux, the ld.so.conf.d file is unnecessary
to find them.

This leaves only libpmi.so in $(libdir)/flux, which will
be outside of the default system library search path,
fixing issue #530.

libpmi.so will be found when $(libdir)/flux is prepended
to LD_LIBRAY_PATH by the broker when it starts the initial
program.  From there it will be inherited by other programs
launched from flux, including another flux instance.