test: add flux_rpc_completed() tests
After testing the fatal error function, reset the fatality bit
so flux_rpc_completed() works properly.

Check that flux_rpc_completed() is true when one of the
RPC's has failed, or when a handle fatal error has occurred,
as discussed in issue #517.

Update message handlers to the new prototypes to avoid warning.