libflux/handle: add flux_fatality()
Add a mechanism for detecting when the flux_t handle has encountered
a fatal error and should no longer be used, even if the user did not
register a fatal error handler function.

This will be used in flux_rpc_completed() to address issue #517

Also, allocate handle with malloc() not zmalloc().  We are already
zeroing it with memset, and which zmalloc were we linking with anyway?