connectors/local: improve broken pipe errors
When the local connector needs to deliver a response or event
message to a client that has disconnected, it logs a "Broken pipe"
error that is not too helpful, as described in #542.

Make the eroror message a bit more helpful by including the topic
string and message type of the message that is being dropped.