broker: add persist-filesystem, persist-directory
A user may set either 'persist-filesystem' or 'persist-directory'
to indicate a location where output from the session should be
preserved on rank 0.

If 'persist-filesystem' is set but not 'persist-directory', the
latter will be created with a unique name under the former.

If neither is set, leave them that way.

The content backing store modules include logic to write their
data to 'persist-directory' if defined, and not remove it at
session termination.  If 'persist-directory' is not defined,
they use 'scratch-directory' and clean up.

Drop
  --sid NAME
Use instead
  --setattr session-id=NAME

Also, when the session-id is derived from simple PMI, drop the
"simple-" prefix because it adds no useful information.

broker: define persist-* attrs always