testsuite: kvs-watch-until: ignore nil result
Do not pass nil result from kvs_watch() to user provided handler.
Most handlers will not expect a nil result, and code in the callback
to handle this case will be messy and inefficient (and likely useless).

After this fix, expected behavior is that "v" parameter in
kvs-watch-until.lua callback shall never be nil.

Fixes #549