broker/content-cache: respect flush-batch-limit
When a backing store is loaded on rank 0, avoid letting the number
of outstanding requests to the backing store exceed the configured
'content-flush-batch-limit'.

The limit was being checked when store requests to the backing store
were begun as a result of a flush request.  It was not being checked
when they were started as a result of new content landing in the cache.

Checking in this case avoids "handle thrashing" when each new message
received by the broker has to be checked against a long list of
outstanding request messages.