bindings/lua: reason string for reactor_stop_error()
Allow an optional "reason" string to be passed to f:reactor_stop_error(),
e.g.

  f:reactor_stop_error ("Fatal timeout")

When Lua f:reactor() returns to the user on error (i.e. rc < 0), this
string will be passed back in return values as (nil, reason) instead
of the default error string, which is strerror (errno).

The reason string is currently saved in the handle as "lua::reason".