modules/content-sophia: improve shutdown logging
When the module is shutting down due to being unloaded, ensure
that there is a debug log at the beginning of shutdown handling
and at the end, after all content has been returned to the cache.

This will enable us to determine if for some reason this is stalling
by examining broker log timestamps.

Debug for #570.