broker/content-cache: fix HWM calc for cache flush
We restart the cache flush in the store response handler
when the number of dirty entries is nonzero and the number
of outstanding flushes is one-half of the high water mark.
To avoid repeatedly walking the cache when nothing can be done,
we must also check that the number of dirty entries minus the
number of outstanding flushes is nonzero.  Only _then_ walk the cache.

I'm going to tentatively say that this fixes #570, though I
may be jumping to conclusions.