connectors/shmem: fix flux_recv() assertion failure
When flux_recv() was called with FLUX_O_NONBLOCK from
a comms module, the shmem connector would assert here:

shmem.c:143: op_recv: Assertion `n == 1' failed.

This was due to incorrect handling of the zmq_poll()
return value, which returns 0 when no events are pending,
not -1 with errno set to EWOULDBLOCK as previously assumed.