broker/module: handle requests after shutdown
To facilitate comms modules that implement distributed
shutdown sequences, module service routes remains in place
until the module exits.  This creates a window for requests
that arrive after the shutdown request but before the module
exits to be ignored.

Add some code to the module thread to drain the socket
of requests after mod_main() exits, sending them ENOSYS.