t/scripts/waitfile.lua: add plugin support
Add support for a "plugin" to be registered with the waitfile.lua script
at runtime. The plugin can be used to add extra callbacks to the flux
reactor in the flux handle being used for waitfile's nominal duties.

This support will offer a more robust internal testing of waitfile.lua,
but is flexible enough for other use cases as well.