broker: add module_set_status_cb()
Replace module_set_rmmod_cb() with module_set_status_cb(), a callback
which tracks module state transitions.  Handle the transition to EXITED
in this callback, as opposed to handling it where the keepalive message
is received and decoded.

This is a bit cleaner and prepares for the broker to take action on
other module state transitions.