broker: handle module transition from INIT
Defer response to insmod until module transitions out of INIT state.
If it transitions directly to EXITED state, capture errno and
return it in the insmod response.

Add module_push_insmod(), module_pop_insmod() to mirror rmmod functions.

Add module_set_errnum(), module_get_errnum() so that errnum returned
in keepalive at transition to EXITED can be captured.

If mod_main() returns -1 but errno is zero, set it to ECONNRESET
as we are returning only the errnum value in the keepalive, not a
separate success/failure flag, and need to capture the fact that
there was an error.