broker: add runlevel class
Create a runlevel abstraction for the broker.
There are three hardwired levels:
1 - initialization
2 - initial program
3 - finalization

A read-only broker attribute 'init.run-level' (rank 0 only)
reflects the current run level.

A read-write broker attribute 'init.mode' (rank 0 only)
reflects the current run mode, "normal" for the sequence
described below, "none" to skip run levels 1 and 3.
Other run modes may be added in the future

Each run level has a subprocess on rank 0 associated with it.
Once the TBON wireup is complete, run level 1 is entered and
its subprocess executes.  Although the subprocess executes on
rank 0, it may load modules or flux-exec on other ranks to
initialize the instance.  Its stdout and stderr are redirected
to the flux log (at LOG_INFO and LOG_ERR respectively).
If the level 1 subprocess terminates by exiting with a zero exit code,
level 2 is entered;  if it terminates abnormally, level 3 is entered.

When run level 2 is entered, the level 2 subprocess (the initial
program) is started.  The instance will remain in level 2 during
normal operation.  The level 2 subprocess inherits stdio from the
rank 0 broker.  When the initial program exits, run level 3 is entered.

When run level 3 is entered, the level 3 subprocess is started.
Concurrently, the shutdown timer is started.  At this time the
shutdown timer remains the primary exit path for the broker,
but once we've cleaned up the broker and module exit paths, we
should be able to use the shutdown timer as a fail safe for
level 3.  The level 3 subprocess stdout and stderr are logged
the same as level 1.

Set in the environment for rc subprocesses:
  FLUX_RUNLEVEL_MASK   the nodeset entering runlevel
This prepares the way for future grow and shrink by providing
a mechanism to run rc scripts with effect restricted to a subset
of nodes being added/removed from the instance.

The rc1,rc3 scripts are located in descending precedence order:

1) environment:  $FLUX_RC1_PATH, $FLUX_RC3_PATH
2) config file:  general.rc1_path, general.rc3_path
3) compiled-in:  ${sysconfdir}/flux/rc1, ${sysconfdir}/flux/rc3

If FLUX_RC1_PATH or FLUX_RC3_PATH are set to the empty string,
this effectively disables subprocess execution at that runlevel.