broker: don't exit if shutdown has not expired
The shutdown code currently always calls the shutdown callback
in the broker with 'expired' true.  However, the design documented
in shutdown.h says it is called on shutdown initiation with 'expired'
false so that the broker can begin an orderly shutdown, a thing it
currently does not know how to do.

Restore the logic for not exiting if 'expired' is false and leave
it at that for now, recognizing that more work is needed for orderly
instance shutdown - see issue #25.