libflux/info: add flux_get_nodeset()
const char *flux_get_nodeset (flux_t h, const char *nodeset,
                              const char *exclude);

This function can be shared among utilities to unify the way that
a --rank argument is handled.

First a nodeset is creawted from the 'nodeset' argument:
If "all", parse the value of the $FLUX_NODESET_ALL environment
variable, if set; otherwise use [0-($size-1)].
If "self", use [$rank].  Otherwise, parse the stringified nodeset.

Next, if the 'exclude' argument is non-NULL, subtract members of
that nodeset from the first one.

Finally, if $FLUX_NODESET_MASK is set, eliminate ranks in the original
set that are not present in the mask.