modules/content-sqlite: compress large blobs
Blobs > 256 bytes will be compressed before being stored in the
database using LZO compression.  A new 'size' integer column is added
to store the uncompressed size of the blob so that the right size
buffer can be allocated for decompression.  If size is -1, this
indicates that the blob is not compressed.

This might help a bit with issue #609