modules/kvs: add kvs_copy(), kvs_move()
kvs_copy() recursively copies a key from one name to another.
If the source key is a directory, the destination key is removed,
then the directory is copied recursively.

kvs_copy() does not follow symlinks in the source; it recreates
them with the same (fully qualified) targets in the destination.

kvs_move() calls kvs_copy(), then kvs_unlink()s the source key.