t/scripts/waitfile.lua: fix race in initial stat()
There was a race in waitfile.lua script because the initial stat(2) on
the file of interest was done before the flux statwatcher was started,
even though the self:check() of the the filewatcher object was called
after the start of the stat watcher.

This fix moves the stat(2) initialization to after statwatcher creation,
ensuring that the file cannot change/appear/etc. between statwatcher's
initial stat, and the first manual check done by the script.

Resolves #612.