libflux/dispatch: fix dispatch reference counting
We need to decrement the dispatcher reference count at the time that
we set the 'destroyed' flag on a message handler, not at the time
that it is freed.

Otherwise if message handlers are destroyed outside of the handle
watcher callback, when the dispatcher is destroyed (e.g. by aux hash
destruction when the flux_t handle is destroyed), its usecount won't
be zero and the code in the destroy function to free deleted handlers,
which would have brought the usecount to zero, won't ever be executed.

This caused memory leaks  such as the following in comms modules:

==16833== 9,770 (32 direct, 9,738 indirect) bytes in 2 blocks are
definitely lost in loss record 609 of 622
==16833==    at 0x4C2DB95: calloc (in
/usr/lib/valgrind/vgpreload_memcheck-amd64-linux.so)
==16833==    by 0x4E647DD: xzmalloc (xzmalloc.c:39)
==16833==    by 0x4E47F31: flux_reactor_create (reactor.c:89)
==16833==    by 0x4E47FE6: flux_get_reactor (reactor.c:113)
==16833==    by 0x40E60D: modservice_register (modservice.c:246)
==16833==    by 0x40C7F2: module_thread (module.c:131)
==16833==    by 0x57FD6A9: start_thread (pthread_create.c:333)
==16833==    by 0x5F26E9C: clone (clone.S:109)
==16833==