modules/wreck: turn off heap trace in wrexec child
Disable heap tracing, if enabled, in the child immediately after
wrexec forks.  If this is not done, a trace dump is triggered from
the child when it exits (I believe).  The trace dump file name has
the same number as the next one dumped by the main broker program.

This is only observed when flux wreckrun runs something, not when
flux exec does.