modules/live: add barrier during initialization
Now that modules are loaded via the rc1 script instead of
in the broker before message routing starts, there is a
race in live module initialization.  The hello message requires
that upstream instances be loaded in order to accurately
determine topology.  Fix this by adding a barrier.

If --barrier-count is set to something other than zero,
live will execute a barrier before sending its hello message.
This ensures that upstream instances are loaded and ready
to receive this message before we send it, so that it doesn't
inadvertantly get handled by another rank upstream.

Set --barrier-count in rc1 when loading live module.

In t0010-generic-utils, load the barrier module first (now a
prerequisite), then load live with --barrier-count.

Fixes #637.