libflux/message: improve robustness of wire proto
Add flux_msg_encode_size() to determine the size of the buffer needed
to encode 'msg', and change flux_msg_encode to take a buffer rather
than allocating one internally.  Leave flux_msg_decode() signature as
is, but replace call to zmsg_decode() with internal implementation.
We no longer dependent on czmq's zmsg_encode(), zmsg_decode().

flux_msg_sendfd() now can send a message in one write(2) call
if the OS allows it to.  flux_msg_recvfd() now can receive a message
in two read(2) calls if the OS allows it to.

A fixed 4K buffer in the 'iobuf' argument is now used for encoding
and decoding messages if they fit; otherwise, fall back to the old
behavior of allocating a temp buffer for that.

Add a magic cookie value to the wire protocol to improve message
framing robustness.

Update message unit tests.