lua: JSON: allow 0 to be detected as integer
lua_tonumber() returns 0 on failure, but 0 is also a valid number.
Thus, do not use the return code of this function when testing for integer
vs floating point number.

Instead, rely on LUA_TNUMBER type to assume lua_tonumber() success,
and switch to modf(3) to test for pure int vs float.