cmd/flux-start: add --pmix support
If flux-start --pmix option is given, bootstrap brokers using PMIx;
that is, flux-start itself implements a PMIx server rather than the
simple PMIv1 wire protocol.

The broker will attempt to boot with its pmi-client library, which only
groks PMIv1 currently.  However, since libpmix.so exports a PMIv1 interface
implemented in terms of PMIx, all we need is for pmi-client to notice
some PMIX_ environment variables, and dlopen libpmix.so instead of
libpmi.so in order to make use of this.

Register a pmix listener callback, and wire a Flux fd watcher to
the flux-start's event loop to accept client connections.
Start the watcher when the first client is created; stop it when
the last client is destroyed to allow normal reactor exit.
This disables the libpmix listneer thread.

Add experimental --pmix-socketpair option for initializing a PMIx
client session using a pair of sockets created with socketpair()
instead of forcing it to connect to a UNIX domain socket.
This requires PMIx_server_register_client_fd() support, proposed
in pmix/master#93.

PMIx support is compiled out if libpmix is not found by configure.