subprocess: support stackable set of hooks
Allow registration of a stack of hooks or callbacks at various points
in the life cycle of a subprocess to extend an modify the libsubprocess
interface to process execution.

The hook function prototype is standardized to take only a pointer
to a struct subprocess. Any context required in the hook must be pushed
into the subprocess via a subprocess_set_context() and retrieved via
subprocess_get_context().

The poorly named `subprocess_set_callback()` is removed as part of this
change, replaced with the  SUBPROCESS_COMPLETE hook insertion point.

Other hook points can be found in the subprocess_hook_t enumeration
in subprocess.h.