tracing and profiling interface and basic backends
This is a *very* rough cut on tracing for flux.  If TAU_PROFILING is
enabled, it turns on some macros that will register flux event
processing with TAU.  Getting flux to build, link, and run with TAU is
another matter entirely.   I recommend building your own TAU with only
pthread support, then specifying tau_cc.sh as the compiler with only
the options to use shared linking and track pthreads while only
affecting linking.  Then when running preload the pthread indirection
library (this should not be necessary, but it is).

Alternately, there is a caliper backend to the basic macros as well,
which can be turned on with the `--with-caliper` option on configure.
One still has to have caliper available in appropriate paths, but it's
much easier to get started with this.  Unfortunately the data is rather
harder to wrangle from this right now.

Either way, the main point of this is to serve as a starting point for
tracing/profiling in flux rather than a full setup.