cmd/flux: rework env/config usage
Give flux(1) the exclusive role of detecting whether flux is running
in installed or in-tree modes, and combining compiled-in paths with
incoming environment to determine outgoing environment.
The flux(1) outgoing environment contains the entire configuration,
which can then be taken up by the broker and mirrored in broker
attributes.

Flux components that have a broker handle can access the broker
attribitues with flux_attr_get().  Sub-commands are also free
to use environment the variables set by flux(1) directly.

Search paths that are modified by flux(1) are de-duplicated so
environment modification should be more or less idempotent,
and running flux(1) within the environment of a Flux instance
should not result in duplicate entries.

Drop flux --uri and --trace options and anticipate that FLUX_URI
or FLUX_HANDLE_TRACE will be set in the inbound environment
if those settings need to be influenced.  Drop --config option
since there is no longer a config file.

Update the following subcommands:
flux-env:
    Use environment rather than flux_conf to print the outgoing
    environment settings.

flux-help:
    Assume MANPATH is already set to include flux man pages.

flux-keygen:
    Add --secdir option to compensate for loss of flux(1) --secdir
    option, needed in tests.

flux-module:
    Assume FLUX_MODULE_PATH is set in the environment.
    Don't fallback to compiled-in default.

Update tests for new flux(1) options:
    t/t0001-basic
    t/t1005-cmddriver.t
    t/t0001-basic.t
    t/t2002-pmi