libflux/conf: make accessor for compiled-in paths
Demote flux_conf class to a mere accessor for compiled-in paths.
If flux_conf_get flags argument is zero, install paths are returned.
If flags is CONF_FLAG_INTREE, paths based on source/build tree are
returned.

Drop the in-tree config file (and support for ZPL config files
generally), as the in-tree config file reflected only the compiled-in
paths set by autoconf.

Drop the special compiled-in paths sprinkled around Makefile.am's,
and have users call the accessor.  In this commit, primarily
tests that behaved that way are updated, and also the fall-back
connector path is dropped from libflux/handle.c.