modules/wreck: add lwj-complete dir in kvs
Add a lwj-complete directory in kvs for ordered list of completed
jobs. As each lwj is "archived" and moved from lwj-active.<id> to
lwj.<id>, link to this directory from lwj-completed.<hb>.<id> where
"hb" is the current Flux heartbeat epoch.

This allows completed jobs to be traversed in order with minimal
kvs overhead.