cmd/help: do not run man on multiple topics
For `flux help <name>`, the flux-help builtin currently attemps
to invoke 'man flux-<name> <name>', so that a user will be displayed
either a matching flux-<name> flux command, or potentially a
`<name>` flux related man page such as API calls, etc.

Unfortunately, this means that if there is a manual page for a
non-Flux command with `<name>`, then after flux-<name> page is
displayed, a second, unrelated page will be shown to the user, causing
annoyance and confusion.

This patch replaces the current behavior with a heuristic that assumes
if the topic upon which the user is requesting help does not start
with "flux", then 'flux-' should probably be prepended. Otherwise,
we simply pass the argument directly to man(1), assuming that the
user is requesting help for one of the API calls or a 'flux-*' command
directly.

Fixes #676