libflux/rpc: rework for matchtag groups
Use matchtag groups instead of blocks.

When groups are used, the unused non-group bits contain the
destination nodeid, which allows us to dispense with the
nodemap array.

The above change limits nodeids used as targets of flux_rpc_multi()
to 1<<20.  If that becomes a problem in the future, we can use sequence
numbers and a mapping array to allow individual nodeids to exceed 1<<20,
within a nodeset size limit of 1<<20.

flux_rpc_multi() now honors the FLUX_RPC_NORESPONSE flag.