libflux: stop watcher on bad return from periodics reschedule
If a periodic watcher "reschedule callback" returns a time in the
past, catch the error and *safely* stop the watcher by scheduling
an ev_prepare callback to call flux_watcher_stop().

This avoids hitting an assertion in libev, and it isn't clear if
there is any other correct approach other than stopping the watcher.