travis-ci: force libev to use clock_gettime()
For testing with libfaketime, libev cannot be allowed to use
the syscall interface for clock_gettime(). If clock_gettime()
is found in librt, libev will try anything possible to avoid
linking with -lrt, so we have to force it by defining

 EV_USE_CLOCK_SYSCALL=0

and

 EV_USE_MONOTONIC=1

So that clock_gettime() is used for realtime and monotonic time.