modules/wreck: fix MPIRUN_RANK in mvapich plugin
MPIRUN_RANK was inadvertently set to local taskid instead of global
taskid in the lua.d/mvapich.lua "plugin". This is likely due to poor
naming choice for the members of the `wreck` object passed to these
plugins, however, for now just update MPIRUN_RANK to be set to
wreck.globalid instead of wreck.taskid.

Fixes #704