libpmi-server: add barrier completion call
Restructure the barrier callback.  If there isn't one,
barriers complete internally.  If there is one, it is called
once when all local processes have entered.

Once the barrier is complete, pmi_simple_server_barrier_complete()
should be called.

This allows the barrier completion to be asynchronous with
respect to barrier entry, which is needed for integrating
the pmi server code into a reactor loop, as discussed in #706.

Rename the barrier callback to barrier_enter() and add an rc
paramter so it can fail.

Update users.