wreck: save mapping of cores to node ranks
During startup, wrexecd has to traverse list of "cores per node"
(or tasks per node) from kvs in order to determine the global rank
number of the first task on the current node.

Save the mapping of nodeids to # of tasks for future use in
the prog_ctx of the wrexecd.