libsubprocess/zio: remove ancient zmq socket arg
The zio stuff in libsubprocess was originally designed to connect
zeromq sockets together for remote and local I/O. However, that
use case was never realized, so now there is this `dstsock` argument
for "reader" zio objects that is completely unused.

This patch removes the dstsock argument, zio member, and associated
code to simplify zio a bit.