broker: avoid deprecated pmi_get_id() call
Drop code in broker PMI bootstrap that set the session-id
attribute based on PMI_Get_id(), if implemented.  This is not
the right function - as documented in RFC 13, it should
simply be an alias for PMI_KVS_Get_my_name().

Enable the fallback code that sets the session-id to
the value returned from PMI_Get_appnum(), which actually
isn't the right thing either, but we will need to come up
with a good design for passing job ID's to children before
we fix that one.