libpmi/simple_client: relax proto parse, cleanup
Use keyval to parse PMI message.

Make rc and msg optional attributes, and allow key=val
tuples to appear out of order.

If rc is returned in a response with a nonzero value,
assume this is a PMI error code and return it as the
function return code.

Avoid calling oom() or xzmalloc(), xstrdup(), etc..