libpmi/simple_server: relax proto parse, cleanup
Parse protocol lines using the keyval class.

Respond to all commands, including unimplemented.

We don't handle spawn yet, but spawn is implemented as a multi
line command (mcmd), and we at least need to be able to ingest
one of those to respond with an appropriate error that allows the
client to continue, rather than simply dropping the connection
with EPROTO.

Instead of triggering a fatal protocol for every little thing
that goes wrong, start returning appropriate PMI errors in the
response if a key or value is too long, or a kvs_put/kvs_get fails.

If the barrier_enter() method returns < 0, exit the
barrier with rc=PMI_FAIL.

Avoid callign oop(), xzmalloc(), xstrdup(), etc