caliper: message tracing and child trace avoidance
This update adds tracing for various properties of messages, RPCs,
events, etc.  This will not give us all we need to get
matches for all messages, but it's on the way.

Caliper is currently pulled in using pkg-config to configure our
build, if that becomes annoying we may want to rework that at some
point.

As an FYI, be aware that the profiling_msg_snapshot routine can be
called due to messaging activity inside connector_init, so the profiling
object is not ready, and potentially cause uninitialized nodes to be
emitted.  This is currently addressed by a check and return in the
snapshot routine, but if at some point those code-paths become threaded
something fancier may be required.