Initial support for AddressSanitizer (ASan) and ThreadSanitizer (TSan).
Add `--enable-sanitizer` as a configure option.

Support ASan and TSan for now. Not LeakSanitizer (LSan) and MemorySanitizer (MSan).
This is because ASan is a superset of LSan and the lower versions
of the compilers available on LLNL systems do not yet support LSan and MSan.

Integration is straightforward. It appeared that Sanitizers do not like
the `--no-undefined` linker option. So, this commit drops this option
for linking when `--enable-sanitizer` is given.

Other follow-up issues have been created in GitHub:
Issue #720, #721, #722, #724 and #725.