resource-hwloc: add walk_topology option to reload request
Require JSON payload in resource-hwloc.reload request with a single
potential member "walk_topology" -- a boolean that, when set, will
replace the current value of ctx->walk_topology during the reload
operation.

This can be used to "reload" topology with broken down values in kvs
after a session is running.