modules/kvs: factor out json_dirent.c
Move the functions for creating a JSON "dirent" and appending it to
an array of "operations" from kvs.c to their own module so they can
be shared.

This prepares for sending all "put" data with the commit request,
instead of as seprate synchronous RPC's.  To do that we need to build
an ordered list of (key, dirent) pairs in a JSON array.

These functions assert on failure, which is not what we want.  However,
fixing this would result in many lines of code change in the KVS, and
since the plan is to change the format to RFC 11 tree objects, it did
not seem like right time to address this.