modules/kvs: send 'put' data with 'commit' request
Consolidate the "put" and "commit" request messages in the client
code, so that the commit request includes an ordered list of
(key, dirent) pairs representing put, unlink, symlink, etc. calls,
which now return immediately, unless a large value is stored.

If kvs_put() receives a value whose JSON encoding is longer than
the size of its blobref, the value is synchronously written to the
content store, and its blobref is referenced in the dirent.

Note that the format of the commit message did not need to change,
since "internal" commits already carried the array of (key, dirent)
operations - they were just empty when the handle code sent a commit.

In addition to (probably) improving performance since many KVS put
operations will now return immediately, this avoids a race where
the new asynchronous fence calls are used, and a put is sent before
an in-progress fence (on the same handle) has completed.

Fixes #729.