modules/kvs: add kvs_fence_set_context()
If kvs_fence_set_context() has been called with the name of an
upcoming fence operation, all subsequent operations (put, unlink,
mkdir, symlink) are queued for that fence, until kvs_fence_clear_context()
is called.  Operations not queued when a fence context is set are queued
anonymously, as before.

A kvs_commit() call flushes out the handle's pending anonymous operations.

A kvs_fence() flushes out only the pending operations that match
its name.  If no operations match its name, then it flushes anonymous
operations, as before.

Consistency note:  operations associated with a fence may complete
out of order with respect to other operations on the handle.

API note:  it probably would be better in the future to implement a
commit handle abstraction, where a commit is opened, operations are
applied to a particular commit, then the commit handle can be passed
to either commit or fence.  For now, though, we just need to support
asynchronous fences co-mingled with synchronous commits and want to
perturb the API as little as possible.

Partially fixes #729