modules/wreck: wrexecd PMI sets KVS fence context
When wrexecd is running a PMI server, it uses asynchronous
kvs_fence().  At the same time, it may be handling standard
I/O output and writing it to the KVS.

Have the PMI server call kvs_fence_set_context () so that the
fence only applies to PMI_KVS_Put() data, not the stdio or other data.

This solves a potential problem.  If the following sequence occurs:
-  stdio data is kvs_put()
-  PMI kvs_fence_begin()
-  stdio kvs_commit()

the kvs_commit() is expected to return only after stdio data
is committed to the kvs namespace, but since the fence ran
in between, the stdio data is not applied until the PMI fence completes.
Since libkz applies a sequence number to each KVS key it "appends"
this likely is resolved quickly without I/O sequencing issues; however,
it is egregiously incorrect KVS behavior.  (Note that this behavior
was introduced with async fence in ae2eb13 three weeks ago).

Partially fixes #729