modules/kvs: split commit handler into master/slave
The function for handling commit reuests is too long.
Split it into one version for master, one for slaves
to improve clarity despite minor code duplication.

Since objects are commited to the content store by clients
now, the scan for dirty references in the commit can
be removed.  This was also the one place where the commit
handler could be stalled/restarted, so eliminate that code.