tracing: link caliper stubs rather than impl.
Flux now links to the stubs by default, and no longer has references to
the pkg-config variables outside of the configuration script.  In order
to use the tracing, build as before and LD_PRELOAD libcaliper.so, with
that it will work as normal, otherwise all of the calls are stubbed out.