libpmi: control export of flux_pmi_library symbol
The flux_pmi_library symbol should only be defined in the libpmi.so
dynamic library.  Its existance is a sanity check that the "wrap"
path in that library is not dlopen()ing itself, which would lead
to an infinite regress in PMI_Init().

Avoid leaking flux_pmi_library via libflux.so.

Fixes #736 (hopefully)